Move anonymous error to object translation from Azure specific to gateway
